create ResetDialogPreferenceFragmentCompat
