+++ collect_app/src/main/java/org/odk/collect/android/formentry/audit/IdentifyUserPromptDialogFragment.java
@@ -22,7 +22,7 @@ public class IdentifyUserPromptDialogFragment extends MaterialFullScreenDialogFr
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.identify_user_dialog, container, false);
+            return inflater.inflate(R.layout.identify_user_dialog, container, false);
     }
 
     @Override
